requirements.txt
jsonpath_ng
google_auth_oauthlib
google-api-python-client
google-auth-httplib2
google-auth
json